Fix bug in my reforge plot changes (r15199).
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@15202 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_reforge_plot.cpp
@@ -215,7 +215,7 @@ void reforge_plot_t::analyze_stats()
     for ( size_t j = 0; j < stat_mods[ i ].size(); j++ )
     {
       stat_e stat = reforge_plot_stat_indices[ j ];
-      double mod = stat_mods[ i ][ j ];
+      int mod = stat_mods[ i ][ j ];
       current_reforge_sim -> enchant.add_stat( stat, mod );
       delta_result[ j ].value = mod;
       delta_result[ j ].error = 0;
